+++ tests/Authorization/Middleware/CheckAdminActivityForTest.php
@@ -28,9 +28,9 @@ class CheckAdminActivityForTest extends TestCase
         $request->route('activity_slug')->willReturn($activity);
         $authentication = $this->prophesize(Authentication::class);
 
-        $user = $this->newUser(['id' => 1]);
-        $group = $this->newGroup(['id' => 5]);
-        $role = $this->newRole(['id' => 10]);
+        $user = $this->newUser();
+        $group = $this->newGroup();
+        $role = $this->newRole();
         $authentication->getUser()->willReturn($user);
         $authentication->getGroup()->willReturn($group);
         $authentication->getRole()->willReturn($role);
@@ -50,9 +50,9 @@ class CheckAdminActivityForTest extends TestCase
         $request->route('test_callback_is_called')->shouldBeCalled()->willReturn(true);
 
         $authentication = $this->prophesize(Authentication::class);
-        $user = $this->newUser(['id' => 1]);
-        $group = $this->newGroup(['id' => 5]);
-        $role = $this->newRole(['id' => 10]);
+        $user = $this->newUser();
+        $group = $this->newGroup();
+        $role = $this->newRole();
         $authentication->getUser()->willReturn($user);
         $authentication->getGroup()->willReturn($group);
         $authentication->getRole()->willReturn($role);